tools/requirements.txt
requests==2.31.0
boto3==1.29.0
python-dotenv==1.0.0
